+++ test/config/test_configmanager.cc
@@ -62,9 +62,14 @@ public:
             grbJs / "playlists.js",
             grbJs / "metadata.js",
             gerberaDir / "sqlite3.sql",
+            gerberaDir / "sqlite3-drop.sql",
             gerberaDir / "sqlite3-upgrade.xml",
             gerberaDir / "mysql.sql",
+            gerberaDir / "mysql-drop.sql",
             gerberaDir / "mysql-upgrade.xml",
+            gerberaDir / "postgres.sql",
+            gerberaDir / "postgres-drop.sql",
+            gerberaDir / "postgres-upgrade.xml",
         };
         std::ofstream file;
         for (auto&& mFile : mockFiles) {